third_party/WebKit/LayoutTests/fast/forms/file/file-input-capture-expected.txt

Tests the capture attribute of <input type='file'>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL 'capture' in input should be true. Was false.
FAIL input.capture should be  (of type string). Was undefined (of type undefined).
FAIL input.capture should be filesystem (of type string). Was undefined (of type undefined).
FAIL input.capture should be camera (of type string). Was undefined (of type undefined).
FAIL input.capture should be camcorder (of type string). Was undefined (of type undefined).
FAIL input.capture should be microphone (of type string). Was undefined (of type undefined).
FAIL input.capture should be filesystem (of type string). Was undefined (of type undefined).
FAIL input.capture should be camcorder. Was CamCorder.
FAIL input.getAttribute('capture') should be CamCorder. Was xyzzy.
PASS successfullyParsed is true

TEST COMPLETE
